# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestión de inventario de camisetas de fútbol y clientes B2B. Proveedor mayorista de camisetas nacionales e internacionales.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "ce1b06de9bbd15a5cc0ed7ab28ee5713",
                "responses": {
                    "200": {
                        "description": "Listado de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cliente"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un cliente",
                "operationId": "703b9106594ca621e0c5bb1b8328bb70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente con sus camisetas",
                "operationId": "9414678b6b3a1f03b3d296d4dcf8d844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente",
                "operationId": "27af8d44cc1dbf176f91e6118f0ce961",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente",
                "description": "No se puede eliminar si tiene camisetas asociadas",
                "operationId": "9ad82979f62d7065a69ed5067066d569",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "409": {
                        "description": "Cliente tiene camisetas asociadas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClienteInput": {
                "required": [
                    "nombre_comercial",
                    "rut",
                    "direccion",
                    "tipo",
                    "contacto_nombre",
                    "contacto_email"
                ],
                "properties": {
                    "nombre_comercial": {
                        "type": "string",
                        "example": "90minutos"
                    },
                    "rut": {
                        "type": "string",
                        "example": "76.111.222-3"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Providencia, Santiago"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@90minutos.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    }
                },
                "type": "object"
            },
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre_comercial": {
                        "type": "string",
                        "example": "90minutos"
                    },
                    "rut": {
                        "type": "string",
                        "example": "76.111.222-3"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Providencia, Santiago"
                    },
                    "tipo": {
                        "type": "string",
                        "example": "Preferencial",
                        "enum": [
                            "Regular",
                            "Preferencial"
                        ]
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "contacto_email": {
                        "type": "string",
                        "example": "juan@90minutos.cl"
                    },
                    "porcentaje_oferta": {
                        "type": "number",
                        "format": "float",
                        "example": 15
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}